avoid format mismatch warnings

+++ libtiff/mkg3states.c
@@ -348,7 +348,7 @@ WriteTable(FILE* fd, const TIFFFaxTabEnt* T, int Size, const char* name)
 	sep = "\n";
 	for (i = 0; i < Size; i++) {
 	    fprintf(fd, "%s%s%d,%d,%d%s",
-		sep, prebrace, T->State, T->Width, T->Param, postbrace);
+		sep, prebrace, T->State, T->Width, (int) T->Param, postbrace);
 	    if (((i+1) % 12) == 0)
 		    sep = ",\n";
 	    else
@@ -359,7 +359,7 @@ WriteTable(FILE* fd, const TIFFFaxTabEnt* T, int Size, const char* name)
 	sep = "\n ";
 	for (i = 0; i < Size; i++) {
 	    fprintf(fd, "%s%s%3d,%3d,%4d%s",
-		sep, prebrace, T->State, T->Width, T->Param, postbrace);
+		sep, prebrace, T->State, T->Width, (int) T->Param, postbrace);
 	    if (((i+1) % 6) == 0)
 		    sep = ",\n ";
 	    else
